requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiolimiter==1.2.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bert-score==0.3.13
bitsandbytes==0.42.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cohere==5.17.0
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastavro==1.12.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keybert==0.9.0
kiwisolver==1.4.8
kmedoids==0.5.3.1
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.15
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
narwhals==1.45.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.0
openai==1.93.0
openTSNE==1.0.2
orjson==3.11.2
packaging==25.0
pandas==2.3.0
pillow==11.3.0
plotly==6.2.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.0
scikit-learn-extra==0.3.0
scipy==1.15.3
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
together==1.5.21
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0
typer==0.15.4
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
voyageai==0.3.4
yarl==1.20.1
zstandard==0.24.0
